Add translations to NFT alt text
