+++ public/locales/pt/common.json
@@ -4,6 +4,8 @@
   "searchId": "Pesquisar Developer DAO id",
   "loading": "Carregando...",
   "error": "Erro",
+  "developerTraits": "Características do desenvolvedor",
+  "failedLoadTraits": "Falha ao carregar características",
   "madeBy": "Feito pela comunidade Developer DAO: ",
   "owner": "Dono",
   "unclaimed": "Não reclamado",
